+++ react/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @webspatial/react-sdk
 
+## 0.1.15
+
+### Patch Changes
+
+- 5b72246: export dts for window.xrCurrentSceneDefaults
+  - @webspatial/core-sdk@0.1.15
+
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies [a87c8bf]
+- Updated dependencies [1a9e18c]
+  - @webspatial/core-sdk@0.1.14
+
 ## 0.1.13
 
 ### Patch Changes